changed skill level field to float

+++ player/models.py
@@ -4,7 +4,7 @@ from django.utils import timezone
 # Create your models here.
 class Player_Attributes(models.Model):
     username = models.CharField(max_length=100)
-    skill_level = models.DecimalField(max_digits=2, decimal_places=2)
+    skill_level = models.FloatField()
     address = models.CharField(max_length=500)
     availability_days = models.CharField(max_length=500)
     availability_times = models.CharField(max_length=500)
